feat: add demand-order page & root purchase-internal endpoint
